refactor: improve project structure, modularize configuration, and resolve dependency cycles
- Extract all unit tests into dedicated external test files in compliance with the 400-line modularity lint.
- Split monolithic config/logic.rs into focused submodules (bandwidth, bittorrent, hooks, general, network, storage, vpn).
- Move resource mapping configuration into config/resource_mapping.rs to eliminate circular dependency between orchestrator and config.
- Fix read-after-write consistency bug in StorageEngine by implementing immediate dirty buffer flushes prior to read operations.
- Enable clippy::cognitive_complexity and clippy::type_complexity lints globally and configure thresholds via clippy.toml.
- Add comprehensive config deserialization tests covering empty, partial, and malformed inputs.
