+++ aura-core/src/config/mod.rs
@@ -1,3 +1,12 @@
+pub mod bandwidth;
+pub mod bittorrent;
 pub mod credentials;
+pub mod general;
+pub mod hooks;
+pub mod network;
+pub mod resource_mapping;
+pub mod storage;
+pub mod vpn;
+
 pub mod logic;
 pub use logic::*;
